[CDF-22685] 😼 Builder Concept Part 1 (#1017)
* build; ignore folder

* refactor; added builders package

* refactor: setup shell for builder

* refactor: moved all code

* refactor; moved the rest

* refactor: switch over

* tests: updated tets

* refactor; cleanup double use

* tests: changed tests

* refactor; fix removed needle folder
