+++ .gitignore
@@ -304,3 +304,4 @@ my_org
 # Org used by bootcamp
 ice-cream-dataops
 module_upgrade/*.log
+needle/
