+++ cognite_toolkit/_cdf_tk/commands/build.py
@@ -1,24 +1,18 @@
 from __future__ import annotations
 
 import contextlib
-import copy
-import difflib
-import re
 import shutil
-import traceback
 from collections import Counter, defaultdict
-from collections.abc import Callable, Hashable, Iterable
+from collections.abc import Hashable, Iterable
 from dataclasses import dataclass, field
-from functools import partial
 from pathlib import Path
 from typing import Any, cast
 
-import yaml
 from rich import print
 from rich.panel import Panel
 from rich.progress import track
 
-from cognite_toolkit._cdf_tk._parameters import ParameterSpecSet
+from cognite_toolkit._cdf_tk.builders import Builder, create_builder
 from cognite_toolkit._cdf_tk.cdf_toml import CDFToml
 from cognite_toolkit._cdf_tk.client import ToolkitClient
 from cognite_toolkit._cdf_tk.commands._base import ToolkitCommand
@@ -26,79 +20,41 @@ from cognite_toolkit._cdf_tk.constants import (
     _RUNNING_IN_BROWSER,
     INDEX_PATTERN,
     ROOT_MODULES,
-    TEMPLATE_VARS_FILE_SUFFIXES,
 )
 from cognite_toolkit._cdf_tk.data_classes import (
     BuildConfigYAML,
     BuildVariables,
     BuiltModule,
     BuiltModuleList,
-    BuiltResource,
     BuiltResourceList,
     ModuleDirectories,
     ModuleLocation,
-    SourceLocationEager,
     SourceLocationLazy,
 )
 from cognite_toolkit._cdf_tk.exceptions import (
-    AmbiguousResourceFileError,
     ToolkitDuplicatedModuleError,
     ToolkitEnvError,
     ToolkitError,
-    ToolkitFileExistsError,
     ToolkitMissingModuleError,
-    ToolkitNotADirectoryError,
-    ToolkitYAMLFormatError,
 )
 from cognite_toolkit._cdf_tk.hints import ModuleDefinition, verify_module_directory
 from cognite_toolkit._cdf_tk.loaders import (
-    LOADER_BY_FOLDER_NAME,
     ContainerLoader,
     DataModelLoader,
-    FileLoader,
-    FileMetadataLoader,
-    FunctionLoader,
-    GroupLoader,
-    Loader,
     NodeLoader,
-    RawDatabaseLoader,
-    RawTableLoader,
     ResourceLoader,
     SpaceLoader,
     ViewLoader,
 )
-from cognite_toolkit._cdf_tk.loaders.data_classes import RawDatabaseTable
 from cognite_toolkit._cdf_tk.tk_warnings import (
-    FileReadWarning,
-    HighSeverityWarning,
     LowSeverityWarning,
-    MediumSeverityWarning,
     MissingDependencyWarning,
-    ToolkitBugWarning,
-    ToolkitNotSupportedWarning,
-    UnresolvedVariableWarning,
-    WarningList,
-)
-from cognite_toolkit._cdf_tk.tk_warnings.fileread import (
-    DuplicatedItemWarning,
-    MissingRequiredIdentifierWarning,
-    UnknownResourceTypeWarning,
 )
 from cognite_toolkit._cdf_tk.utils import (
     CDFToolConfig,
-    calculate_str_or_file_hash,
-    humanize_collection,
-    module_from_path,
-    quote_int_value_by_key_in_yaml,
-    read_yaml_content,
-    resource_folder_from_path,
-    safe_read,
-    safe_write,
 )
 from cognite_toolkit._cdf_tk.validation import (
-    validate_data_set_is_set,
     validate_modules_variables,
-    validate_resource_yaml,
 )
 from cognite_toolkit._version import __version__
 
@@ -111,6 +67,7 @@ class BuildCommand(ToolkitCommand):
 
         # Built State
         self._module_names_by_variable_key: dict[str, list[str]] = defaultdict(list)
+        self._builder_by_resource_folder: dict[str, Builder] = {}
         self._state = _BuildState()
         self._has_built = False
 
@@ -176,7 +133,7 @@ class BuildCommand(ToolkitCommand):
         verbose: bool = False,
         ToolGlobals: CDFToolConfig | None = None,
         progress_bar: bool = False,
-    ) -> tuple[BuiltModuleList, dict[Path, Path]]:
+    ) -> BuiltModuleList:
         is_populated = build_dir.exists() and any(build_dir.iterdir())
         if is_populated and clean:
             shutil.rmtree(build_dir)
@@ -220,6 +177,7 @@ class BuildCommand(ToolkitCommand):
 
         # Setup state before building modules
         self._module_names_by_variable_key.clear()
+        self._builder_by_resource_folder.clear()
         for variable in variables:
             for module_location in modules:
                 if variable.location in module_location.relative_path.parts:
@@ -238,7 +196,7 @@ class BuildCommand(ToolkitCommand):
         build_environment.dump_to_file(build_dir)
         if not _RUNNING_IN_BROWSER:
             self.console(f"Build complete. Files are located in {build_dir!s}/")
-        return built_modules, self._state.source_by_build_path
+        return built_modules
 
     def build_modules(
         self,
@@ -304,68 +262,19 @@ class BuildCommand(ToolkitCommand):
             )
 
         for resource_name, resource_files in module.source_paths_by_resource_folder.items():
-            build_plugin = {
-                FileMetadataLoader.folder_name: partial(self._expand_file_metadata, module=module, verbose=verbose),
-            }.get(resource_name)
-
-            built_resource_list = BuiltResourceList[Hashable]()
-            for source_path in resource_files:
-                if source_path.suffix.lower() not in TEMPLATE_VARS_FILE_SUFFIXES or self._is_exception_file(
-                    source_path, resource_name
-                ):
-                    continue
-
-                destination = self._state.create_destination_path(source_path, resource_name, module.dir, build_dir)
-
-                built_resources = self._build_resources(
-                    source_path, destination, module_variables, build_plugin, verbose
+            if resource_name not in self._builder_by_resource_folder:
+                self._builder_by_resource_folder[resource_name] = create_builder(
+                    resource_name, build_dir, self._module_names_by_variable_key, silent=self.silent, verbose=verbose
                 )
+            builder = self._builder_by_resource_folder[resource_name]
 
-                built_resource_list.extend(built_resources)
-
-            if resource_name == FunctionLoader.folder_name:
-                self._validate_function_directory(built_resource_list, module=module)
-                self.copy_function_directory_to_build(built_resource_list, module.dir, build_dir)
+            built_resource_list = builder.build_resource_folder(resource_files, module_variables, module)
 
             build_resources_by_folder[resource_name] = built_resource_list
+
+            self.warning_list.extend(builder.last_build_warnings())
+
         return build_resources_by_folder
-
-    def _build_resources(
-        self,
-        source_path: Path,
-        destination_path: Path,
-        variables: BuildVariables,
-        build_plugin: Callable[[str], str] | None,
-        verbose: bool,
-    ) -> BuiltResourceList:
-        if verbose:
-            self.console(f"Processing {source_path.name}")
-
-        destination_path.parent.mkdir(parents=True, exist_ok=True)
-
-        content = safe_read(source_path)
-
-        location = SourceLocationEager(source_path, calculate_str_or_file_hash(content, shorten=True))
-        # Todo: Remove when state is removed
-        self._state.source_by_build_path[destination_path] = source_path
-
-        content = variables.replace(content, source_path.suffix)
-        if build_plugin is not None and source_path.suffix.lower() in {".yaml", ".yml"}:
-            content = build_plugin(content)
-
-        safe_write(destination_path, content)
-
-        file_warnings, identifiers_kind_pairs = self.validate(content, source_path, destination_path)
-
-        if file_warnings:
-            self.warning_list.extend(file_warnings)
-            # Here we do not use the self.warn method as we want to print the warnings as a group.
-            if self.print_warning:
-                print(str(file_warnings))
-
-        return BuiltResourceList(
-            [BuiltResource(identifier, location, kind, destination_path) for identifier, kind in identifiers_kind_pairs]
-        )
 
     @staticmethod
     def _validate_modules(
@@ -414,28 +323,6 @@ class BuildCommand(ToolkitCommand):
                 f"the environment ({config.environment.name})?"
             )
 
-    def _validate_function_directory(self, built_resources: BuiltResourceList, module: ModuleLocation) -> None:
-        has_config_files = any(resource.kind == FunctionLoader.kind for resource in built_resources)
-        if has_config_files:
-            return
-        config_files_misplaced = [
-            file
-            for file in module.source_paths_by_resource_folder[FunctionLoader.folder_name]
-            if FunctionLoader.is_supported_file(file)
-        ]
-        if config_files_misplaced:
-            for yaml_source_path in config_files_misplaced:
-                required_location = module.dir / FunctionLoader.folder_name / yaml_source_path.name
-                self.warn(
-                    LowSeverityWarning(
-                        f"The required Function resource configuration file "
-                        f"was not found in {required_location.as_posix()!r}. "
-                        f"The file {yaml_source_path.as_posix()!r} is currently "
-                        f"considered part of the Function's artifacts and "
-                        f"will not be processed by the Toolkit."
-                    )
-                )
-
     def _check_missing_dependencies(self, project_config_dir: Path, ToolGlobals: CDFToolConfig | None = None) -> None:
         existing = {
             (resource_cls, id_) for resource_cls, ids in self._state.ids_by_resource_type.items() for id_ in ids
@@ -478,269 +365,6 @@ class BuildCommand(ToolkitCommand):
         ):
             return True
         return False
-
-    @staticmethod
-    def _is_exception_file(filepath: Path, resource_directory: str) -> bool:
-        # In the 'functions' resource directories, all `.yaml` files must be in the root of the directory
-        # This is to allow for function code to include arbitrary yaml files.
-        # In addition, all files in not int the 'functions' directory are considered other files.
-        return resource_directory == FunctionLoader.folder_name and filepath.parent.name != FunctionLoader.folder_name
-
-    def copy_function_directory_to_build(
-        self,
-        built_resources: BuiltResourceList,
-        module_dir: Path,
-        build_dir: Path,
-    ) -> None:
-        function_directory_by_name = {
-            dir_.name: dir_ for dir_ in (module_dir / FunctionLoader.folder_name).iterdir() if dir_.is_dir()
-        }
-        external_id_by_function_path = self._read_function_path_by_external_id(
-            built_resources, function_directory_by_name
-        )
-
-        for external_id, function_path in external_id_by_function_path.items():
-            # Function directory already validated to exist in read function
-            function_dir = function_directory_by_name[external_id]
-            destination = build_dir / FunctionLoader.folder_name / external_id
-            if destination.exists():
-                raise ToolkitFileExistsError(
-                    f"Function {external_id!r} is duplicated. If this is unexpected, ensure you have a clean build directory."
-                )
-            shutil.copytree(function_dir, destination)
-
-            # Clean up cache files
-            for subdir in destination.iterdir():
-                if subdir.is_dir():
-                    shutil.rmtree(subdir / "__pycache__", ignore_errors=True)
-            shutil.rmtree(destination / "__pycache__", ignore_errors=True)
-
-    def _read_function_path_by_external_id(
-        self, built_resources: BuiltResourceList, function_directory_by_name: dict[str, Path]
-    ) -> dict[str, str | None]:
-        function_path_by_external_id: dict[str, str | None] = {}
-        for built_resource in built_resources:
-            if built_resource.kind != FunctionLoader.kind or built_resource.destination is None:
-                continue
-            source_file = built_resource.destination
-            try:
-                raw_content = read_yaml_content(safe_read(source_file))
-            except yaml.YAMLError as e:
-                raise ToolkitYAMLFormatError(f"Failed to load function files {source_file.as_posix()!r} due to: {e}")
-            raw_functions = raw_content if isinstance(raw_content, list) else [raw_content]
-            for raw_function in raw_functions:
-                external_id = raw_function.get("externalId")
-                function_path = raw_function.get("functionPath")
-                if not external_id:
-                    self.warn(
-                        HighSeverityWarning(
-                            f"Function in {source_file.as_posix()!r} has no externalId defined. "
-                            f"This is used to match the function to the function directory."
-                        )
-                    )
-                    continue
-                elif external_id not in function_directory_by_name:
-                    raise ToolkitNotADirectoryError(
-                        f"Function directory not found for externalId {external_id} defined in {source_file.as_posix()!r}."
-                    )
-                if not function_path:
-                    self.warn(
-                        MediumSeverityWarning(
-                            f"Function {external_id} in {source_file.as_posix()!r} has no function_path defined."
-                        )
-                    )
-                function_path_by_external_id[external_id] = function_path
-
-        return function_path_by_external_id
-
-    def _expand_file_metadata(self, raw_content: str, module: ModuleLocation, verbose: bool) -> str:
-        try:
-            raw_list = read_yaml_content(raw_content)
-        except yaml.YAMLError as e:
-            raise ToolkitYAMLFormatError(f"Failed to load file definitions file {module.dir} due to: {e}")
-
-        is_file_template = (
-            isinstance(raw_list, list)
-            and len(raw_list) == 1
-            and FileMetadataLoader.template_pattern in raw_list[0].get("externalId", "")
-        )
-        if not is_file_template:
-            return raw_content
-        if not (isinstance(raw_list, list) and raw_list and isinstance(raw_list[0], dict)):
-            raise ToolkitYAMLFormatError(
-                f"Expected a list with a single dictionary in the file metadata file {module.dir}, "
-                f"but got {type(raw_list).__name__}"
-            )
-        template = raw_list[0]
-        if verbose:
-            self.console(
-                f"Detected file template name {FileMetadataLoader.template_pattern!r} in {module.relative_path.as_posix()!r}"
-                f"Expanding file metadata..."
-            )
-        expanded_metadata: list[dict[str, Any]] = []
-        for filepath in module.source_paths_by_resource_folder[FileLoader.folder_name]:
-            if not FileLoader.is_supported_file(filepath):
-                continue
-            new_entry = copy.deepcopy(template)
-            new_entry["externalId"] = new_entry["externalId"].replace(
-                FileMetadataLoader.template_pattern, filepath.name
-            )
-            new_entry["name"] = filepath.name
-            expanded_metadata.append(new_entry)
-        return yaml.safe_dump(expanded_metadata)
-
-    def validate(
-        self,
-        content: str,
-        source_path: Path,
-        destination: Path,
-    ) -> tuple[WarningList[FileReadWarning], list[tuple[Hashable, str]]]:
-        warning_list = WarningList[FileReadWarning]()
-        module = module_from_path(source_path)
-        resource_folder = resource_folder_from_path(source_path)
-
-        all_unmatched = re.findall(pattern=r"\{\{.*?\}\}", string=content)
-        for unmatched in all_unmatched:
-            warning_list.append(UnresolvedVariableWarning(source_path, unmatched))
-            variable = unmatched[2:-2]
-            if module_names := self._module_names_by_variable_key.get(variable):
-                module_str = (
-                    f"{module_names[0]!r}"
-                    if len(module_names) == 1
-                    else (", ".join(module_names[:-1]) + f" or {module_names[-1]}")
-                )
-                self.console(
-                    f"The variables in 'config.[ENV].yaml' need to be organised in a tree structure following"
-                    f"\n    the folder structure of the modules, but can also be moved up the config hierarchy to be shared between modules."
-                    f"\n    The variable {variable!r} is defined in the variable section{'s' if len(module_names) > 1 else ''} {module_str}."
-                    f"\n    Check that {'these paths reflect' if len(module_names) > 1 else 'this path reflects'} the location of {module}.",
-                    prefix="    [bold green]Hint:[/] ",
-                )
-
-        if destination.suffix not in {".yaml", ".yml"}:
-            return warning_list, []
-
-        # Ensure that all keys that are version gets read as strings.
-        # This is required by DataModels, Views, and Transformations that reference DataModels and Views.
-        content = quote_int_value_by_key_in_yaml(content, key="version")
-        try:
-            parsed = yaml.safe_load(content)
-        except yaml.YAMLError as e:
-            if self.print_warning:
-                print(str(warning_list))
-            raise ToolkitYAMLFormatError(
-                f"YAML validation error for {destination.name} after substituting config variables: {e}"
-            )
-
-        loader = self._get_loader(resource_folder, destination, source_path)
-        if loader is None or not issubclass(loader, ResourceLoader):
-            return warning_list, []
-
-        api_spec = self._get_api_spec(loader, destination)
-        is_dict_item = isinstance(parsed, dict)
-        items = [parsed] if is_dict_item else parsed
-
-        identifier_kind_pairs: list[tuple[Hashable, str]] = []
-        for no, item in enumerate(items, 1):
-            element_no = None if is_dict_item else no
-
-            identifier: Any | None = None
-            # Raw Tables and Raw Databases can have different loaders in the same file.
-            item_loader = loader
-            try:
-                identifier = item_loader.get_id(item)
-            except KeyError as error:
-                if loader is RawTableLoader:
-                    try:
-                        identifier = RawDatabaseLoader.get_id(item)
-                        item_loader = RawDatabaseLoader
-                    except KeyError:
-                        warning_list.append(
-                            MissingRequiredIdentifierWarning(source_path, element_no, tuple(), error.args)
-                        )
-                else:
-                    warning_list.append(MissingRequiredIdentifierWarning(source_path, element_no, tuple(), error.args))
-
-            if identifier:
-                if item_loader is RawTableLoader:
-                    database = RawDatabaseTable(identifier.db_name)
-                    if database not in self._state.ids_by_resource_type[RawDatabaseLoader]:
-                        self._state.ids_by_resource_type[RawDatabaseLoader][database] = source_path
-
-                identifier_kind_pairs.append((identifier, item_loader.kind))
-                if first_seen := self._state.ids_by_resource_type[item_loader].get(identifier):
-                    warning_list.append(DuplicatedItemWarning(source_path, identifier, first_seen))
-                else:
-                    self._state.ids_by_resource_type[item_loader][identifier] = source_path
-
-                for dependency in loader.get_dependent_items(item):
-                    self._state.dependencies_by_required[dependency].append((identifier, source_path))
-
-            if api_spec is not None:
-                resource_warnings = validate_resource_yaml(parsed, api_spec, source_path, element_no)
-                warning_list.extend(resource_warnings)
-
-            data_set_warnings = validate_data_set_is_set(items, loader.resource_cls, source_path)
-            warning_list.extend(data_set_warnings)
-
-        return warning_list, identifier_kind_pairs
-
-    def _get_api_spec(self, loader: type[ResourceLoader], destination: Path) -> ParameterSpecSet | None:
-        api_spec: ParameterSpecSet | None = None
-        try:
-            api_spec = loader.get_write_cls_parameter_spec()
-        except Exception as e:
-            # Todo Replace with an automatic message to sentry.
-            self.warn(
-                ToolkitBugWarning(
-                    header=f"Failed to validate {destination.name} due to: {e}", traceback=traceback.format_exc()
-                )
-            )
-        return api_spec
-
-    def _get_loader(self, resource_folder: str, destination: Path, source_path: Path) -> type[Loader] | None:
-        folder_loaders = LOADER_BY_FOLDER_NAME.get(resource_folder, [])
-        if not folder_loaders:
-            self.warn(
-                ToolkitNotSupportedWarning(
-                    f"resource of type {resource_folder!r} in {source_path.name}.",
-                    details=f"Available resources are: {', '.join(LOADER_BY_FOLDER_NAME.keys())}",
-                )
-            )
-            return None
-
-        loaders = [loader for loader in folder_loaders if loader.is_supported_file(destination)]
-        if len(loaders) == 0:
-            suggestion: str | None = None
-            if "." in source_path.stem:
-                core, kind = source_path.stem.rsplit(".", 1)
-                match = difflib.get_close_matches(kind, [loader.kind for loader in folder_loaders])
-                if match:
-                    suggested_name = f"{core}.{match[0]}{source_path.suffix}"
-                    suggestion = f"Did you mean to call the file {suggested_name!r}?"
-            else:
-                kinds = [loader.kind for loader in folder_loaders]
-                if len(kinds) == 1:
-                    suggestion = f"Did you mean to call the file '{source_path.stem}.{kinds[0]}{source_path.suffix}'?"
-                else:
-                    suggestion = f"All files in the {resource_folder!r} folder must have a file extension that matches the resource type. Supported types are: {humanize_collection(kinds)}."
-            self.warn(UnknownResourceTypeWarning(source_path, suggestion))
-            return None
-        elif len(loaders) > 1 and all(loader.folder_name == "raw" for loader in loaders):
-            # Multiple raw loaders load from the same file.
-            return RawTableLoader
-        elif len(loaders) > 1 and all(issubclass(loader, GroupLoader) for loader in loaders):
-            # There are two group loaders, one for resource scoped and one for all scoped.
-            return GroupLoader
-        elif len(loaders) > 1:
-            names = " or ".join(f"{destination.stem}.{loader.kind}{destination.suffix}" for loader in loaders)
-            raise AmbiguousResourceFileError(
-                f"Ambiguous resource file {destination.name} in {destination.parent.name} folder. "
-                f"Unclear whether it is {' or '.join(loader.kind for loader in loaders)}."
-                f"\nPlease name the file {names}."
-            )
-
-        return loaders[0]
 
 
 @dataclass
