--- needle/config.dev.yaml
@@ -1,12 +1,0 @@
-environment:
-  name: dev
-  project: needle
-  type: dev
-  selected:
-  - modules/
-
-variables:
-  modules:
-    common:
-      cdf_auth_readwrite_all:
-        readwrite_source_id: da20b44a-236a-4e6e-9f46-68f741df2315
