--- needle/modules/common/cdf_auth_readwrite_all/module.toml
@@ -1,8 +1,0 @@
-[module]
-title = "Admin group for CI/CD"
-
-[packages]
-tags = [
-    "inrobot",
-    "quickstart"
-]
